implement request handling with wren
