+++ examples/quarks.wren
@@ -1,0 +1,13 @@
+class QuarksAPI {
+  foreign static getheader( ctx, x )
+  foreign static request( x )
+  foreign static get( x )
+  foreign static getkeys( x, skip, limit )
+  foreign static put( x )
+  foreign static incrval( x )
+}
+
+class QuarksEnv {
+  foreign static request( x )
+}
+
